+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v4.16.0
   - Release December 19th 2023
-  - New `:color(scope, type)` formatter only for DOCX template: `{d.myColor:color(p)}`
+  - Fix: accept undefined `formatOptions` when `convertTo` is an object
+  - [EE] New `:color(scope, type)` formatter only for DOCX template: `{d.myColor:color(p)}`
     - scope can be `p`  (by default), `cell`, `row` for applying color on the current paragraph, cell and row.
     - type can be `text` (by default), `background` for cells and rows.